Move the M-RoPE layouts to modeling_multimodal_utils
`configuration_utils` imports `modeling_rope_utils` for
`RotaryEmbeddingConfigMixin`, so that module has to stay importable in a
torch-free install — which is why every torch annotation in it is
deferred. Hosting 800+ lines of tensor code there was a standing trap:
CI caught it once already when a `dtype: torch.dtype = torch.long`
default was evaluated at import time.

The layouts belong with the mixins that call them anyway:
`modeling_multimodal_utils` imports torch unconditionally and is only
reached from modeling code. Move `get_mrope_index`, the four layouts and
their helpers, `get_mrope_vision_positions` / `get_mrope_image_positions`
/ `get_mrope_text_positions` and `uses_mrope` there, leaving
`modeling_rope_utils` as the frequency machinery it was.

Splits the tests the same way: `tests/utils/test_modeling_multimodal_utils.py`
holds the layout tests, `test_modeling_rope_utils.py` the frequency ones.
