+++ src/transformers/modeling_rope_utils.py
@@ -14,7 +14,6 @@
 
 from __future__ import annotations
 
-import itertools
 import math
 import warnings
 from collections.abc import Callable
@@ -1084,828 +1083,3 @@ def rope_config_validation(config: RotaryEmbeddingConfigMixin, ignore_keys: set 
     )
     config.standardize_rope_params()
     config.validate_rope()
-
-
-# ── Multimodal (M-RoPE) decoder positions ────────────────────────────────────
-# Where a multimodal model's decoder position ids come from: text tokens keep 1D positions while each
-# image/video span gets its own axes. `get_mrope_index` is the entry point, `_MROPE_LAYOUT_FUNCTIONS` the
-# layouts; a model declares which one it uses (and the knobs it needs) on its config.
-
-
-def get_mrope_vision_positions(
-    start_position: int,
-    grid_thw: list[int] | torch.Tensor,
-    temporal_merge_size: int = 1,
-    spatial_merge_size: int = 1,
-    time_interval: float = 1,
-    dtype: torch.dtype | None = None,
-    device: str | torch.device | None = None,
-) -> torch.Tensor:
-    """3D M-RoPE positions (temporal, height, width) for one image/video grid, in the *decoder* sequence.
-
-    This is the decoder-side counterpart to [`get_vision_position_ids`] (which produces the vision
-    *encoder*'s rotary positions): it lays out a single grid's `(T, H, W)` patches as three position axes,
-    each offset by `start_position` (the running position in the token sequence). The temporal axis is
-    additionally scaled by `time_interval` (video temporal spacing). Merge sizes downscale the grid the way
-    the vision backbone does. Pure function of its arguments — the model passes its config-derived spec.
-
-    Args:
-        start_position: running position offset of this grid in the decoder sequence.
-        grid_thw: `(3,)` `(T, H, W)` grid of the image/video after patch embedding.
-        temporal_merge_size: temporal backbone merge factor (`T // temporal_merge_size`).
-        spatial_merge_size: spatial backbone merge factor (`H, W // spatial_merge_size`).
-        time_interval: spacing between consecutive temporal position ids. May be fractional, in which case
-            the scaled temporal ids are truncated to `dtype`.
-        dtype: dtype of the returned positions.
-        device: device for the returned tensor.
-
-    Returns:
-        `(3, T'*H'*W')` — temporal/height/width position ids, offset by `start_position`.
-    """
-    dtype = dtype if dtype is not None else torch.long
-    llm_grid_t = grid_thw[0].item() // temporal_merge_size
-    llm_grid_h = grid_thw[1].item() // spatial_merge_size
-    llm_grid_w = grid_thw[2].item() // spatial_merge_size
-    position_temporal = (torch.arange(llm_grid_t, device=device) * time_interval).to(dtype)
-    position_height = torch.arange(llm_grid_h, dtype=dtype, device=device) + start_position
-    position_width = torch.arange(llm_grid_w, dtype=dtype, device=device) + start_position
-    t_grid, h_grid, w_grid = torch.meshgrid(position_temporal, position_height, position_width, indexing="ij")
-    vision_position_ids = torch.stack([t_grid, h_grid, w_grid], dim=0).reshape(3, -1)
-    vision_position_ids[0] += start_position  # temporal offset, after the time_interval scaling
-    return vision_position_ids
-
-
-def uses_mrope(config) -> bool:
-    """Whether `config` declares M-RoPE at all — its text rope parameters carry an `mrope_section` (the
-    per-axis head split only multi-axis models have). `False` for plain decoders and for VLMs that keep 1D
-    text positions (Llava & co.), which is the signal to leave `position_ids` alone."""
-    text_config = config.get_text_config()
-    rope_parameters = getattr(text_config, "rope_parameters", None)
-    if isinstance(rope_parameters, dict) and "mrope_section" in rope_parameters:
-        return True
-    return "mrope_section" in (getattr(text_config, "ignore_keys_at_rope_validation", None) or ())
-
-
-def get_mrope_text_positions(
-    attention_mask: torch.Tensor,
-    num_axes: int = 3,
-    dtype: torch.dtype | None = None,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Plain 1D positions broadcast over `num_axes` M-RoPE axes: `(position_ids, rope_deltas)`.
-
-    What a multi-axis model falls back to when a sequence has no vision span to lay out — every token
-    (including audio ones, which carry no spatial axes of their own) just counts up, padded slots keeping
-    position 1.
-    """
-    dtype = dtype if dtype is not None else torch.long
-    position_ids = attention_mask.to(dtype).cumsum(-1) - 1
-    position_ids.masked_fill_(attention_mask == 0, 1)
-    position_ids = position_ids.unsqueeze(0).expand(num_axes, -1, -1)
-    max_position_ids = position_ids.max(0, keepdim=False)[0].max(-1, keepdim=True)[0]
-    return position_ids, max_position_ids + 1 - attention_mask.sum(dim=-1, keepdim=True)
-
-
-def _mrope_place_positions(
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor | None,
-    num_axes: int,
-    attention_mask: torch.Tensor | None,
-    positions_for_sequence,
-    dtype: torch.dtype | None = None,
-    padded_position: int = 0,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Shared frame every M-RoPE layout sits in: `(position_ids, rope_deltas)`.
-
-    Per batch row it drops the padded tokens, asks `positions_for_sequence(token_ids, token_types)` for that
-    sequence's `(num_axes, seq)` positions — which is where the layouts differ — scatters them back onto the
-    unpadded slots, and records the delta `generate` advances decode positions by. `dtype` defaults to
-    `input_ids`' (a layout that lays out fractional temporal positions asks for a float one) and
-    `padded_position` is what the masked-out slots keep.
-    """
-    position_ids = torch.full(
-        (num_axes, input_ids.shape[0], input_ids.shape[1]),
-        padded_position,
-        dtype=dtype or input_ids.dtype,
-        device=input_ids.device,
-    )
-    rope_deltas = []
-    for batch_idx, token_ids in enumerate(input_ids):
-        token_types = mm_token_type_ids[batch_idx] if mm_token_type_ids is not None else None
-        valid_tokens = None
-        if attention_mask is not None:
-            valid_tokens = attention_mask[batch_idx].bool()
-            token_ids = token_ids[valid_tokens]
-            token_types = token_types[valid_tokens] if token_types is not None else None
-
-        positions = positions_for_sequence(token_ids, token_types)
-        positions = positions.to(device=position_ids.device, dtype=position_ids.dtype)
-
-        if valid_tokens is not None:
-            position_ids[:, batch_idx, valid_tokens] = positions
-        else:
-            position_ids[:, batch_idx] = positions
-        rope_deltas.append(positions.max() + 1 - len(token_ids))
-    return position_ids, torch.tensor(rope_deltas, device=input_ids.device).unsqueeze(1)
-
-
-def _mrope_index_interleaved_runs(
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor,
-    *,
-    spatial_merge_size: int,
-    attention_mask: torch.Tensor | None = None,
-    image_grid_thw: torch.LongTensor | None = None,
-    video_grid_thw: torch.LongTensor | None = None,
-    second_per_grid_ts: torch.Tensor | None = None,
-    tokens_per_second: float | None = None,
-    split_video_frames: bool = False,
-    video_temporal_merge_size: int = 1,
-    **unused,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """3-axis M-RoPE laid out run by run: `(position_ids, mrope_deltas)`.
-
-    Text runs use standard 1D positions (broadcast across the 3 axes); image/video runs use 3D
-    (temporal, height, width) positions from [`get_mrope_vision_positions`]. Runs are delimited by
-    `mm_token_type_ids` (0=text, 1=image, 2=video). A pure function driven entirely by the caller's spec —
-    each VLM's `get_rope_index` becomes a thin wrapper passing its config-derived values, so the layout
-    lives here instead of being copied per model.
-
-    Args:
-        input_ids: `(batch, seq)` token ids.
-        mm_token_type_ids: `(batch, seq)` per-token modality marker (0=text, 1=image, 2=video).
-        spatial_merge_size: spatial backbone merge factor.
-        attention_mask: `(batch, seq)` padding mask; positions are placed on the unpadded tokens.
-        image_grid_thw: `(num_images, 3)` grids, consumed in order for image runs.
-        video_grid_thw: `(num_videos, 3)` grids, consumed in order for video runs.
-        second_per_grid_ts: `(num_videos,)` per-video seconds-per-temporal-grid.
-        tokens_per_second: when set, video temporal spacing scales by `tokens_per_second *
-            second_per_grid_ts`; otherwise consecutive temporal ids are 1 apart.
-        split_video_frames: expand each video grid to one `T=1` row per frame before placement, for
-            processors that separate frames with timestamps.
-        video_temporal_merge_size: temporal backbone merge factor, applied to video grids only.
-
-    Returns:
-        position_ids: `(3, batch, seq)` long.
-        mrope_position_deltas: `(batch, 1)` — `max_position + 1 - unpadded_len` per batch row.
-    """
-    if split_video_frames and video_grid_thw is not None:
-        video_grid_thw = torch.repeat_interleave(video_grid_thw, video_grid_thw[:, 0], dim=0)
-        video_grid_thw[:, 0] = 1
-
-    grid_iters = {
-        1: iter(image_grid_thw) if image_grid_thw is not None else None,
-        2: iter(video_grid_thw) if video_grid_thw is not None else None,
-    }
-    second_per_grid_iter = (
-        iter(second_per_grid_ts) if second_per_grid_ts is not None else iter([1] * input_ids.shape[1])
-    )
-
-    def positions_for_sequence(token_ids, token_types):
-        input_type_group = []
-        for key, group in itertools.groupby(enumerate(token_types.tolist()), lambda x: x[1]):
-            group = list(group)
-            start_index = group[0][0]
-            end_index = group[-1][0] + 1
-            input_type_group.append((key, start_index, end_index))
-
-        current_pos = 0
-        llm_pos_ids_list = []
-        for modality_type, start_idx, end_idx in input_type_group:
-            if modality_type == 0:
-                text_len = end_idx - start_idx
-                llm_pos_ids_list.append(_mrope_positions_block(text_len, current_pos, device=token_ids.device))
-                current_pos += text_len
-            else:
-                grid_thw = next(grid_iters[modality_type])
-                if modality_type == 2 and tokens_per_second is not None:
-                    time_interval = tokens_per_second * int(next(second_per_grid_iter))
-                else:
-                    time_interval = 1
-                vision_position_ids = get_mrope_vision_positions(
-                    current_pos,
-                    grid_thw,
-                    temporal_merge_size=video_temporal_merge_size if modality_type == 2 else 1,
-                    spatial_merge_size=spatial_merge_size,
-                    time_interval=time_interval,
-                    device=token_ids.device,
-                )
-                llm_pos_ids_list.append(vision_position_ids)
-                current_pos += max(grid_thw[1], grid_thw[2]) // spatial_merge_size
-        return torch.cat(llm_pos_ids_list, dim=1).reshape(3, -1)
-
-    return _mrope_place_positions(input_ids, mm_token_type_ids, 3, attention_mask, positions_for_sequence)
-
-
-def _mrope_index_indexed_images(
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor,
-    *,
-    num_axes: int,
-    spatial_merge_size: int,
-    attention_mask: torch.Tensor | None = None,
-    image_grid_thw: torch.LongTensor | None = None,
-    **unused,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """M-RoPE over a per-image ordinal instead of a temporal axis: `(position_ids, rope_deltas)`.
-
-    HunYuanVL's layout. Every token starts from plain 1D positions (a text baseline that is *overwritten*,
-    not accumulated) and each image span replaces its slice with `(width, height, image_index)` on the last
-    three of `mrope_section`'s axes — earlier axes keep the 1D positions. The pooled image grid carries one
-    newline-style token per row, so the width channel spans `w + 1` (see `get_mrope_image_positions`), and a
-    span may include the two image-boundary tokens.
-
-    Returns:
-        position_ids: `(num_axes, batch, seq)` long — unlike the other layouts this is not fixed at 3 axes,
-            since `mrope_section` decides how many there are and only the last 3 carry the image.
-        rope_deltas: `(batch, 1)` — `max_position + 1 - unpadded_len` per batch row.
-    """
-    if num_axes < 3:
-        raise ValueError(f"An indexed-image M-RoPE layout needs at least 3 axes, got {num_axes}.")
-
-    grid_iter = iter(image_grid_thw) if image_grid_thw is not None else None
-    image_index = 0
-
-    def positions_for_sequence(token_ids, token_types):
-        nonlocal image_index
-        current_position_ids = torch.arange(token_ids.shape[-1], dtype=input_ids.dtype, device=token_ids.device)
-        current_position_ids = current_position_ids.view(1, -1).expand(num_axes, -1).clone()
-        if grid_iter is None:
-            return current_position_ids
-
-        for modality_type, group in itertools.groupby(enumerate(token_types.tolist()), lambda x: x[1]):
-            if modality_type != 1:  # image == 1; this layout has no video/audio modality
-                continue
-            group = list(group)
-            span_start, span_end = group[0][0], group[-1][0] + 1
-            try:
-                grid_thw = next(grid_iter)
-            except StopIteration as error:
-                raise ValueError("Found more image placeholder spans than entries in `image_grid_thw`.") from error
-
-            vision_position_ids = get_mrope_image_positions(
-                grid_thw[1:], spatial_merge_size=spatial_merge_size, device=token_ids.device
-            )
-            grid_tokens = vision_position_ids.shape[1]
-            span_length = span_end - span_start
-            if span_length == grid_tokens + 2:  # span includes the image-boundary tokens
-                grid_start = span_start + 1
-            elif span_length == grid_tokens:
-                grid_start = span_start
-            else:
-                raise ValueError(
-                    "Image placeholder span length does not match `image_grid_thw`: "
-                    f"span_length={span_length}, expected {grid_tokens} or {grid_tokens + 2}."
-                )
-
-            grid_end = grid_start + grid_tokens
-            offset = num_axes - 3
-            current_position_ids[offset : offset + 2, grid_start:grid_end] = vision_position_ids.to(
-                dtype=input_ids.dtype
-            )
-            current_position_ids[offset + 2, grid_start:grid_end] = image_index
-            image_index += 1
-        return current_position_ids
-
-    position_ids, rope_deltas = _mrope_place_positions(
-        input_ids, mm_token_type_ids, num_axes, attention_mask, positions_for_sequence
-    )
-    if image_grid_thw is not None and image_index != len(image_grid_thw):
-        raise ValueError(
-            "Number of image placeholder spans does not match `image_grid_thw`: "
-            f"spans={image_index}, images={len(image_grid_thw)}."
-        )
-    return position_ids, rope_deltas
-
-
-def get_mrope_image_positions(
-    grid_hw: list[int] | torch.Tensor,
-    spatial_merge_size: int = 1,
-    device: str | torch.device | None = None,
-) -> torch.Tensor:
-    """`(2, tokens)` `(width, height)` indices for one pooled image grid, for the indexed-image layout.
-
-    The vision merger appends one newline-style token per image row, so the width channel spans `w + 1`
-    positions while the height channel repeats each row id over that extra column.
-    """
-    grid_h, grid_w = (int(value) for value in grid_hw)
-    llm_grid_h = grid_h // spatial_merge_size
-    llm_grid_w = grid_w // spatial_merge_size
-    position_height, position_width = torch.meshgrid(
-        torch.arange(llm_grid_h, dtype=torch.long, device=device),
-        torch.arange(llm_grid_w + 1, dtype=torch.long, device=device),
-        indexing="ij",
-    )
-    return torch.stack([position_width.flatten(), position_height.flatten()], dim=0)
-
-
-def _mrope_positions_block(
-    length: int | torch.Tensor,
-    start_position: int | torch.Tensor,
-    num_axes: int = 3,
-    dtype: torch.dtype | None = None,
-    device: str | torch.device | None = None,
-) -> torch.Tensor:
-    """`(num_axes, length)` block of consecutive positions from `start_position`, the same on every axis.
-
-    What a text run — or an audio span, which is 1D in time — contributes to a multi-axis layout.
-    """
-    dtype = dtype if dtype is not None else torch.long
-    return torch.arange(int(length), dtype=dtype, device=device).view(1, -1).expand(num_axes, -1) + start_position
-
-
-def _mrope_temporal_chunks(
-    temporal_positions: torch.Tensor, positions_per_chunk: int, start_position: int | torch.Tensor
-) -> list[tuple[int, int]]:
-    """`(start, end)` slices cutting a monotonic temporal axis into successive `positions_per_chunk` ranges.
-
-    Used to interleave a video and its own audio track in time: both are cut at the same chunk boundaries
-    (`position_id_per_seconds * seconds_per_chunk` positions ≈ one chunk of wall-clock time), then emitted
-    chunk by chunk. `start_position` is the span's own origin, subtracted before chunking.
-    """
-    chunks = []
-    chunk_start, chunk_index = 0, 1
-    for index in range(len(temporal_positions)):
-        if temporal_positions[index] - start_position >= chunk_index * positions_per_chunk:
-            chunks.append((chunk_start, index))
-            chunk_start = index
-            chunk_index += 1
-    chunks.append((chunk_start, len(temporal_positions)))
-    return chunks
-
-
-def _mrope_index_audio_chunked(
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor | None = None,
-    *,
-    spatial_merge_size: int,
-    image_token_id: int,
-    video_token_id: int,
-    audio_token_id: int,
-    vision_start_token_id: int,
-    audio_start_token_id: int,
-    position_id_per_seconds: float,
-    seconds_per_chunk: float,
-    attention_mask: torch.Tensor | None = None,
-    image_grid_thw: torch.LongTensor | None = None,
-    video_grid_thw: torch.LongTensor | None = None,
-    second_per_grid_ts: torch.Tensor | None = None,
-    audio_seqlens: torch.LongTensor | None = None,
-    use_audio_in_video: bool = False,
-    **unused,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """3-axis M-RoPE over image, video **and audio** spans, audio-in-video interleaved in time chunks.
-
-    Qwen2.5-Omni's layout. Spans are located by scanning `input_ids` for placeholder tokens rather than by
-    grouping `mm_token_type_ids` (which this family does not produce): at each step the *nearest* upcoming
-    image/video/audio placeholder wins, and its span contributes a one-position `bos` block, the modality's
-    own positions, and a one-position `eos` block, each starting one past the previous block's maximum.
-    Audio is 1D in time, images and videos are 3D (see [`get_mrope_vision_positions`]), and the temporal axis
-    counts `position_id_per_seconds` positions per second of media. With `use_audio_in_video`, a video and
-    its soundtrack share one span: both are cut into `seconds_per_chunk` chunks by
-    [`_mrope_temporal_chunks`] and emitted video-chunk-then-audio-chunk, wrapped in doubled bos/eos blocks.
-
-    Args:
-        input_ids: `(batch, seq)` token ids.
-        mm_token_type_ids: unused — spans come from the placeholder token ids.
-        spatial_merge_size: spatial backbone merge factor.
-        image_token_id, video_token_id, audio_token_id: placeholder token ids marking each modality's span.
-        vision_start_token_id, audio_start_token_id: the tokens opening a vision/audio span, used to count
-            the spans in the sequence.
-        position_id_per_seconds: temporal positions per second of media.
-        seconds_per_chunk: audio-in-video interleaving granularity, in seconds.
-        attention_mask: `(batch, seq)` padding mask; positions are placed on the unpadded tokens.
-        image_grid_thw: `(num_images, 3)` grids, consumed in order for image spans.
-        video_grid_thw: `(num_videos, 3)` grids, consumed in order for video spans.
-        second_per_grid_ts: `(num_videos,)` per-video seconds-per-temporal-grid.
-        audio_seqlens: `(num_audios,)` mel-frame lengths, consumed in order for audio spans.
-        use_audio_in_video: lay a video out interleaved with its own audio track.
-
-    Returns:
-        position_ids: `(3, batch, seq)` long, masked-out slots left at 1.
-        mrope_position_deltas: `(batch, 1)` — `max_position + 1 - unpadded_len` per batch row.
-    """
-    image_idx, video_idx, audio_idx = 0, 0, 0
-
-    def positions_for_sequence(token_ids, _token_types):
-        nonlocal image_idx, video_idx, audio_idx
-        device = token_ids.device
-        input_tokens = token_ids.tolist()
-        blocks = []
-
-        def next_position():
-            return blocks[-1].max() + 1 if blocks else 0
-
-        def block(length, start_position=None):
-            start_position = next_position() if start_position is None else start_position
-            return _mrope_positions_block(length, start_position, device=device)
-
-        vision_start_indices = torch.argwhere(token_ids == vision_start_token_id).squeeze(1)
-        vision_tokens = token_ids[vision_start_indices + 1]
-        num_audios = (token_ids == audio_start_token_id).sum()
-        num_images = (vision_tokens == image_token_id).sum()
-        num_videos = (
-            (vision_tokens == audio_start_token_id).sum()
-            if use_audio_in_video
-            else (vision_tokens == video_token_id).sum()
-        )
-        remain_images, remain_videos, remain_audios = num_images, num_videos, num_audios
-        num_spans = num_images + num_audios if use_audio_in_video else num_images + num_videos + num_audios
-
-        st = 0
-        for _ in range(num_spans):
-            unreachable = len(input_tokens) + 1
-            ed_image = (
-                input_tokens.index(image_token_id, st)
-                if image_token_id in input_tokens and remain_images > 0
-                else unreachable
-            )
-            ed_video = (
-                input_tokens.index(video_token_id, st)
-                if video_token_id in input_tokens and remain_videos > 0
-                else unreachable
-            )
-            ed_audio = (
-                input_tokens.index(audio_token_id, st)
-                if audio_token_id in input_tokens and remain_audios > 0
-                else unreachable
-            )
-            min_ed = min(ed_image, ed_video, ed_audio)
-            # the span's own opening token sits between the text run and the placeholders (two of them for
-            # audio-in-video, which opens with both a vision and an audio start token)
-            text_len = min_ed - st - (2 if min_ed == ed_video and use_audio_in_video else 1)
-            if text_len != 0:
-                blocks.append(block(text_len))
-            bos_len = eos_len = 1
-
-            if min_ed == ed_audio:
-                blocks.append(block(bos_len))
-                audio_len = _mrope_audio_length_pooled(audio_seqlens[audio_idx])
-                blocks.append(block(audio_len))
-                blocks.append(block(eos_len))
-                st += int(text_len + bos_len + audio_len + eos_len)
-                audio_idx += 1
-                remain_audios -= 1
-
-            elif min_ed == ed_image:
-                blocks.append(block(bos_len))
-                grid_thw = image_grid_thw[image_idx]
-                blocks.append(
-                    get_mrope_vision_positions(
-                        next_position(),
-                        grid_thw,
-                        spatial_merge_size=spatial_merge_size,
-                        time_interval=position_id_per_seconds,
-                        device=device,
-                    )
-                )
-                image_len = grid_thw.prod() // (spatial_merge_size**2)
-                blocks.append(block(eos_len))
-                st += int(text_len + bos_len + image_len + eos_len)
-                image_idx += 1
-                remain_images -= 1
-
-            elif min_ed == ed_video and not use_audio_in_video:
-                blocks.append(block(bos_len))
-                grid_thw = video_grid_thw[video_idx]
-                blocks.append(
-                    get_mrope_vision_positions(
-                        next_position(),
-                        grid_thw,
-                        spatial_merge_size=spatial_merge_size,
-                        time_interval=float(second_per_grid_ts[video_idx]) * position_id_per_seconds,
-                        device=device,
-                    )
-                )
-                video_len = grid_thw.prod() // (spatial_merge_size**2)
-                blocks.append(block(eos_len))
-                st += int(text_len + bos_len + video_len + eos_len)
-                video_idx += 1
-                remain_videos -= 1
-
-            elif min_ed == ed_video and use_audio_in_video:
-                bos_position = next_position()
-                blocks.append(block(bos_len, bos_position))
-                blocks.append(block(bos_len, bos_position))
-
-                span_start = next_position()
-                grid_thw = video_grid_thw[video_idx]
-                audio_len = _mrope_audio_length_pooled(audio_seqlens[audio_idx])
-                audio_positions = block(audio_len, span_start)
-                video_positions = get_mrope_vision_positions(
-                    span_start,
-                    grid_thw,
-                    spatial_merge_size=spatial_merge_size,
-                    time_interval=float(second_per_grid_ts[video_idx]) * position_id_per_seconds,
-                    device=device,
-                )
-                positions_per_chunk = int(position_id_per_seconds * seconds_per_chunk)
-                video_chunks = _mrope_temporal_chunks(video_positions[0], positions_per_chunk, span_start)
-                audio_chunks = _mrope_temporal_chunks(audio_positions[0], positions_per_chunk, span_start)
-                for video_chunk, audio_chunk in itertools.zip_longest(video_chunks, audio_chunks):
-                    if video_chunk is not None:
-                        blocks.append(video_positions[:, video_chunk[0] : video_chunk[1]])
-                    if audio_chunk is not None:
-                        blocks.append(audio_positions[:, audio_chunk[0] : audio_chunk[1]])
-                video_len = grid_thw.prod() // (spatial_merge_size**2)
-
-                eos_position = next_position()
-                blocks.append(block(eos_len, eos_position))
-                blocks.append(block(eos_len, eos_position))
-                st += int(text_len + 2 * bos_len + audio_len + video_len + 2 * eos_len)
-                audio_idx += 1
-                video_idx += 1
-                remain_videos -= 1
-                remain_audios -= 1
-
-        if st < len(input_tokens):
-            blocks.append(block(len(input_tokens) - st))
-        return torch.cat(blocks, dim=1).reshape(3, -1)
-
-    return _mrope_place_positions(
-        input_ids, mm_token_type_ids, 3, attention_mask, positions_for_sequence, padded_position=1
-    )
-
-
-def _mrope_index_audio_merged(
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor | None = None,
-    *,
-    spatial_merge_size: int,
-    image_token_id: int,
-    video_token_id: int,
-    audio_token_id: int,
-    vision_start_token_id: int,
-    audio_start_token_id: int,
-    position_id_per_seconds: float,
-    audio_window_size: int,
-    attention_mask: torch.Tensor | None = None,
-    image_grid_thw: torch.LongTensor | None = None,
-    video_grid_thw: torch.LongTensor | None = None,
-    second_per_grid_ts: torch.Tensor | None = None,
-    audio_seqlens: torch.LongTensor | None = None,
-    use_audio_in_video: bool = False,
-    **unused,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """3-axis M-RoPE over image, video and audio spans, audio-in-video merged position by position.
-
-    Qwen3-Omni's layout. Like [`_mrope_index_audio_chunked`] it locates spans by scanning `input_ids`, but
-    it scans for the *opening* tokens (`vision_start_token_id` / `audio_start_token_id`) — so the text run
-    ends where the opening token starts, and positions advance by counting emitted tokens instead of
-    re-reading the previous block's maximum. Audio-in-video is merged position by position (whichever of the
-    two streams is earlier in time goes next) rather than in chunks, and its span is wrapped in two-position
-    bos/eos blocks. Positions are float, since a fractional `second_per_grid_ts` leaves the temporal axis
-    fractional here.
-
-    Args:
-        input_ids: `(batch, seq)` token ids.
-        mm_token_type_ids: unused — spans come from the opening token ids.
-        spatial_merge_size: spatial backbone merge factor.
-        image_token_id, video_token_id, audio_token_id: placeholder token ids marking each modality's span.
-        vision_start_token_id, audio_start_token_id: the tokens opening a vision/audio span.
-        position_id_per_seconds: temporal positions per second of media.
-        audio_window_size: audio encoder window (`audio_config.n_window`), which sets how many decoder
-            tokens an audio of a given mel length takes.
-        attention_mask: `(batch, seq)` padding mask; positions are placed on the unpadded tokens.
-        image_grid_thw: `(num_images, 3)` grids, consumed in order for image spans.
-        video_grid_thw: `(num_videos, 3)` grids, consumed in order for video spans.
-        second_per_grid_ts: `(num_videos,)` per-video seconds-per-temporal-grid.
-        audio_seqlens: `(num_audios,)` mel-frame lengths, consumed in order for audio spans.
-        use_audio_in_video: lay a video out interleaved with its own audio track.
-
-    Returns:
-        position_ids: `(3, batch, seq)` float.
-        mrope_position_deltas: `(batch, 1)` — `max_position + 1 - unpadded_len` per batch row.
-    """
-    image_idx, video_idx, audio_idx = 0, 0, 0
-
-    def positions_for_sequence(token_ids, _token_types):
-        nonlocal image_idx, video_idx, audio_idx
-        device = token_ids.device
-        input_tokens = token_ids.tolist()
-        blocks = []
-
-        def block(length, start_position):
-            return _mrope_positions_block(length, start_position, dtype=torch.float, device=device)
-
-        vision_start_indices = torch.argwhere(token_ids == vision_start_token_id).squeeze(1)
-        vision_tokens = token_ids[vision_start_indices + 1]
-        num_audios = (token_ids == audio_start_token_id).sum()
-        num_images = (vision_tokens == image_token_id).sum()
-        num_videos = (
-            (vision_tokens == audio_start_token_id).sum()
-            if use_audio_in_video
-            else (vision_tokens == video_token_id).sum()
-        )
-        remain_images, remain_videos, remain_audios = num_images, num_videos, num_audios
-        num_spans = num_images + num_audios if use_audio_in_video else num_images + num_videos + num_audios
-
-        st = 0
-        for _ in range(num_spans):
-            start_position = blocks[-1].max() + 1 if blocks else 0
-            unreachable = len(input_tokens) + 1
-            ed_vision_start = (
-                input_tokens.index(vision_start_token_id, st)
-                if (image_token_id in input_tokens or video_token_id in input_tokens)
-                and (remain_videos > 0 or remain_images > 0)
-                else unreachable
-            )
-            ed_audio_start = (
-                input_tokens.index(audio_start_token_id, st)
-                if audio_token_id in input_tokens and remain_audios > 0
-                else unreachable
-            )
-            min_ed = min(ed_vision_start, ed_audio_start)
-
-            text_len = min_ed - st
-            if text_len != 0:
-                blocks.append(block(text_len, start_position))
-                start_position += text_len
-
-            audio_in_video = min_ed == ed_vision_start and ed_vision_start + 1 == ed_audio_start
-            bos_len = eos_len = 2 if audio_in_video else 1
-            blocks.append(block(bos_len, start_position))
-            start_position += bos_len
-
-            if min_ed == ed_audio_start:
-                audio_len = _mrope_audio_length_windowed(audio_seqlens[audio_idx], audio_window_size)
-                blocks.append(block(audio_len, start_position))
-                st += int(text_len + bos_len + audio_len + eos_len)
-                audio_idx += 1
-                remain_audios -= 1
-
-            elif min_ed == ed_vision_start and token_ids[ed_vision_start + 1] == image_token_id:
-                grid_thw = image_grid_thw[image_idx]
-                blocks.append(
-                    get_mrope_vision_positions(
-                        start_position,
-                        grid_thw,
-                        spatial_merge_size=spatial_merge_size,
-                        time_interval=position_id_per_seconds,
-                        dtype=torch.float,
-                        device=device,
-                    )
-                )
-                st += int(text_len + bos_len + grid_thw.prod() // (spatial_merge_size**2) + eos_len)
-                image_idx += 1
-                remain_images -= 1
-
-            elif min_ed == ed_vision_start and token_ids[ed_vision_start + 1] == video_token_id:
-                grid_thw = video_grid_thw[video_idx]
-                blocks.append(
-                    get_mrope_vision_positions(
-                        start_position,
-                        grid_thw,
-                        spatial_merge_size=spatial_merge_size,
-                        time_interval=float(second_per_grid_ts[video_idx]) * position_id_per_seconds,
-                        dtype=torch.float,
-                        device=device,
-                    )
-                )
-                st += int(text_len + bos_len + grid_thw.prod() // (spatial_merge_size**2) + eos_len)
-                video_idx += 1
-                remain_videos -= 1
-
-            elif audio_in_video:
-                grid_thw = video_grid_thw[video_idx]
-                audio_len = _mrope_audio_length_windowed(audio_seqlens[audio_idx], audio_window_size)
-                audio_positions = block(audio_len, start_position)
-                video_positions = get_mrope_vision_positions(
-                    start_position,
-                    grid_thw,
-                    spatial_merge_size=spatial_merge_size,
-                    time_interval=float(second_per_grid_ts[video_idx]) * position_id_per_seconds,
-                    dtype=torch.float,
-                    device=device,
-                )
-                # merge the two streams by temporal position, one token at a time, video first on a tie
-                video_pos, audio_pos = 0, 0
-                while video_pos < video_positions.shape[-1] and audio_pos < audio_positions.shape[-1]:
-                    if video_positions[0][video_pos] <= audio_positions[0][audio_pos]:
-                        blocks.append(video_positions[:, video_pos : video_pos + 1])
-                        video_pos += 1
-                    else:
-                        blocks.append(audio_positions[:, audio_pos : audio_pos + 1])
-                        audio_pos += 1
-                if video_pos < video_positions.shape[-1]:
-                    blocks.append(video_positions[:, video_pos:])
-                if audio_pos < audio_positions.shape[-1]:
-                    blocks.append(audio_positions[:, audio_pos:])
-                video_len = grid_thw.prod() // (spatial_merge_size**2)
-                st += int(text_len + bos_len + audio_len + video_len + eos_len)
-                audio_idx += 1
-                video_idx += 1
-                remain_videos -= 1
-                remain_audios -= 1
-
-            blocks.append(block(eos_len, blocks[-1].max() + 1))
-
-        if st < len(input_tokens):
-            blocks.append(block(len(input_tokens) - st, blocks[-1].max() + 1 if blocks else 0))
-        return torch.cat(blocks, dim=1).reshape(3, -1)
-
-    return _mrope_place_positions(
-        input_ids, mm_token_type_ids, 3, attention_mask, positions_for_sequence, dtype=torch.float
-    )
-
-
-def _mrope_audio_length_pooled(audio_seqlen: torch.Tensor | int) -> torch.Tensor | int:
-    """Decoder tokens one audio takes in the chunked layout: two stride-2 convs, then a stride-2 pooler."""
-    return ((audio_seqlen - 1) // 2 + 1 - 2) // 2 + 1
-
-
-def _mrope_audio_length_windowed(audio_seqlen: torch.Tensor | int, window_size: int) -> torch.Tensor | int:
-    """Decoder tokens one audio takes in the merged layout: three stride-2 convs over fixed-size windows,
-    each full `2 * window_size`-frame window collapsing to 13 tokens."""
-    chunk_len = window_size * 2
-    feat_len = (audio_seqlen % chunk_len - 1) // 2 + 1
-    return ((feat_len - 1) // 2 + 1 - 1) // 2 + 1 + (audio_seqlen // chunk_len) * 13
-
-
-# Layout name -> implementation, the way `ACT2FN` maps an activation name. A model declares which one it
-# uses with `config.mrope_layout`; a family whose positions are laid out differently adds its function here
-# (and names it in its config) rather than growing an existing layout. Every layout is a pure function of
-# `(config, inputs)`, which is what lets a model's `get_rope_index` be a one-line wrapper AND lets the
-# exporters compute positions from a saved config with no model instance.
-_MROPE_LAYOUT_FUNCTIONS: dict[str, callable] = {
-    "interleaved_runs": _mrope_index_interleaved_runs,
-    "indexed_images": _mrope_index_indexed_images,
-    "audio_chunked": _mrope_index_audio_chunked,
-    "audio_merged": _mrope_index_audio_merged,
-}
-
-
-def get_mrope_index(
-    config,
-    input_ids: torch.LongTensor,
-    mm_token_type_ids: torch.IntTensor | None = None,
-    *,
-    attention_mask: torch.Tensor | None = None,
-    image_grid_thw: torch.LongTensor | None = None,
-    video_grid_thw: torch.LongTensor | None = None,
-    second_per_grid_ts: torch.Tensor | None = None,
-    audio_seqlens: torch.LongTensor | None = None,
-    use_audio_in_video: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """M-RoPE decoder positions for a `vision + text` sequence: `(position_ids, rope_deltas)`.
-
-    Runs the layout `config.mrope_layout` declares, with that layout's knobs read off the config here — the
-    one place config is turned into M-RoPE parameters, so it works from a saved config with no model
-    instance (a VLM's `get_rope_index` is a one-line wrapper over this, and the exporters drive exported
-    VLMs through the very same call). The layouts themselves (`_MROPE_LAYOUT_FUNCTIONS`) take plain values,
-    so they stay usable and testable on their own.
-
-    Args:
-        config: the model config, declaring `mrope_layout` and its layout's knobs.
-        input_ids: `(batch, seq)` token ids.
-        mm_token_type_ids: `(batch, seq)` per-token modality marker (0=text, 1=image, 2=video); layouts that
-            locate their spans by placeholder token id instead do not need it.
-        image_grid_thw: `(num_images, 3)` grids, consumed in order for image spans.
-        video_grid_thw: `(num_videos, 3)` grids, consumed in order for video spans.
-        attention_mask: `(batch, seq)` padding mask; positions are placed on the unpadded tokens.
-        second_per_grid_ts: `(num_videos,)` per-video seconds-per-temporal-grid.
-        audio_seqlens: `(num_audios,)` mel-frame lengths, consumed in order for audio spans.
-        use_audio_in_video: lay each video out interleaved with its own audio track.
-
-    A layout ignores the tensors it has no use for (an image-only layout takes no video grids).
-
-    Returns:
-        position_ids: `(num_axes, batch, seq)` — 3 axes for every layout except `indexed_images`, which
-            takes its axis count from `mrope_section`; long, except layouts whose temporal axis is
-            fractional (`audio_merged`), which return float.
-        rope_deltas: `(batch, 1)` — `max_position + 1 - unpadded_len` per batch row, in every layout.
-    """
-    layout_name = getattr(config, "mrope_layout", None)
-    layout = _MROPE_LAYOUT_FUNCTIONS.get(layout_name)
-    if layout is None:
-        raise NotImplementedError(
-            f"`{getattr(config, 'model_type', type(config).__name__)}` declares "
-            f"`mrope_layout={layout_name!r}`, which is not implemented here (known layouts: "
-            f"{sorted(_MROPE_LAYOUT_FUNCTIONS)}). Its own `get_rope_index` is the reference: port it to a "
-            "`_MROPE_LAYOUT_FUNCTIONS` entry (a pure function of its inputs) and name it in the config."
-        )
-    vision_config = getattr(config, "vision_config", config)
-    audio_config = getattr(config, "audio_config", config)
-    rope_parameters = getattr(config.get_text_config(), "rope_parameters", None) or {}
-    return layout(
-        input_ids,
-        mm_token_type_ids,
-        spatial_merge_size=vision_config.spatial_merge_size,
-        tokens_per_second=getattr(vision_config, "tokens_per_second", None),
-        video_temporal_merge_size=getattr(vision_config, "temporal_merge_size", None) or 1,
-        split_video_frames=getattr(vision_config, "timestamped_video_frames", False),
-        num_axes=len(rope_parameters.get("mrope_section", [])),
-        image_token_id=getattr(config, "image_token_id", None),
-        video_token_id=getattr(config, "video_token_id", None),
-        audio_token_id=getattr(config, "audio_token_id", None),
-        vision_start_token_id=getattr(config, "vision_start_token_id", None),
-        audio_start_token_id=getattr(config, "audio_start_token_id", None),
-        position_id_per_seconds=getattr(config, "position_id_per_seconds", None),
-        seconds_per_chunk=getattr(config, "seconds_per_chunk", None),
-        audio_window_size=getattr(audio_config, "n_window", None),
-        attention_mask=attention_mask,
-        image_grid_thw=image_grid_thw,
-        video_grid_thw=video_grid_thw,
-        second_per_grid_ts=second_per_grid_ts,
-        audio_seqlens=audio_seqlens,
-        use_audio_in_video=use_audio_in_video,
-    )
