+++ src/transformers/models/hunyuan_vl/modular_hunyuan_vl.py
@@ -29,8 +29,8 @@ from ...generation import GenerationMixin
 from ...image_processing_backends import PilBackend, TorchvisionBackend
 from ...image_utils import PILImageResampling, SizeDict
 from ...masking_utils import create_causal_mask
+from ...modeling_multimodal_utils import get_mrope_index
 from ...modeling_outputs import BaseModelOutputWithPast, BaseModelOutputWithPooling, CausalLMOutputWithPast
-from ...modeling_rope_utils import get_mrope_index
 from ...modeling_utils import ALL_ATTENTION_FUNCTIONS
 from ...processing_utils import Unpack
 from ...utils import TransformersKwargs, auto_docstring, can_return_tuple
