+++ src/transformers/models/qwen2_5_omni/modular_qwen2_5_omni.py
@@ -31,8 +31,9 @@ from ... import initialization as init
 from ...cache_utils import Cache
 from ...configuration_utils import PreTrainedConfig
 from ...generation import GenerationMixin
+from ...modeling_multimodal_utils import get_mrope_index, get_mrope_text_positions
 from ...modeling_outputs import BaseModelOutputWithPooling, CausalLMOutputWithPast, ModelOutput
-from ...modeling_rope_utils import RopeParameters, get_mrope_index, get_mrope_text_positions
+from ...modeling_rope_utils import RopeParameters
 from ...modeling_utils import ALL_ATTENTION_FUNCTIONS, PreTrainedModel
 from ...processing_utils import Unpack
 from ...utils import (
